+++ pestifer/tasks/mdtask.py
@@ -194,10 +194,10 @@ class MDTask(VMDTask):
         cpu_override = specs.get('cpu-override', False)
         logger.debug(f'CPU-override is {cpu_override}')
         na.writescript(params, cpu_override=cpu_override)
-        logger.info(f'Consolidating NAMD parameters for {self.basename} using PSF {state.psf.name}')
+        logger.debug(f'Consolidating CHARMMFF parameters for {self.basename} using PSF {state.psf.name}')
         minimal_prm = na.consolidate_params(state.psf.name)
         if minimal_prm:
-            logger.info(f'Parameter consolidation complete; registering {minimal_prm} as artifact')
+            logger.debug(f'Parameter consolidation complete; registering {minimal_prm} as artifact')
             self.register(minimal_prm, key='charmmff_minimal_prm', artifact_type=CharmmffParFileArtifact)
         else:
             logger.warning(f'Parameter consolidation skipped for {self.basename}; using full parameter set')